requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
allure-pytest==2.9.45
    # via -r requirements.in
allure-python-commons==2.9.45
    # via allure-pytest
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
atomicwrites==1.4.0
    # via pytest
attrs==21.4.0
    # via
    #   allure-python-commons
    #   outcome
    #   pytest
    #   trio
certifi==2022.6.15
    # via urllib3
cffi==1.15.0
    # via
    #   cryptography
    #   trio
colorama==0.4.5
    # via pytest
cryptography==37.0.2
    # via
    #   pyopenssl
    #   urllib3
h11==0.13.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
outcome==1.2.0
    # via trio
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via
    #   allure-python-commons
    #   pytest
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pyopenssl==22.0.0
    # via urllib3
pyparsing==3.0.9
    # via packaging
pysocks==1.7.1
    # via urllib3
pytest==7.1.1
    # via
    #   -r requirements.in
    #   allure-pytest
selenium==4.1.3
    # via -r requirements.in
six==1.16.0
    # via
    #   allure-pytest
    #   allure-python-commons
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
tomli==2.0.1
    # via pytest
trio==0.21.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
urllib3[secure,socks]==1.26.9
    # via selenium
wsproto==1.1.0
    # via trio-websocket
